.astype(float)

+++ server.py
@@ -61,8 +61,8 @@ def ndvi():
     # bottom[bottom == 0] = 0.01
     # ndvi = np.subtract(ir, r)/bottom
 
-    ir = img[:, :, 0]
-    r = img[:, :, 2]
+    ir = img[:, :, 0].astype(float)
+    r = img[:, :, 2].astype(float)
     bottom = np.add(ir, r)
     bottom[bottom == 0] = 0.01
     ndvi = np.subtract(ir, r)/bottom
